refactor plugin.py: move helpers and validators to their own files
